requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
colorama==0.4.4
coverage==6.3.2
crispy-bootstrap5==0.6
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.2
django-autoslug==1.9.8
django-cloudinary-storage==0.3.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.1.1
django-filter==21.1
django-model-utils==4.1.0
factory-boy==3.1.0
Faker==13.3.4
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
ipython==7.19.0
jedi==0.18.1
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
pluggy==0.13.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
py==1.11.0
Pygments==2.11.2
pyparsing==3.0.7
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
traitlets==5.1.1
tzdata==2021.5
urllib3==1.26.9
wcwidth==0.2.5
whitenoise==6.0.0
